Refactor server routes and models into modular structure
