+++ routes/stickerRoutes.js
@@ -1,0 +1,75 @@
+const express = require('express');
+const fetch = require('node-fetch');
+const zlib = require('zlib');
+const { Sticker } = require('../models');
+
+const router = express.Router();
+
+router.get('/sticker/:sticker_id/json', async (req, res) => {
+  try {
+    const sticker = await Sticker.findOne({ file_unique_id: req.params.sticker_id });
+    if (!sticker || !sticker.file_path || !sticker.file_path.endsWith('.tgs')) {
+      return res.status(404).json({ error: 'Sticker not found or not animated' });
+    }
+
+    const telegramUrl = `https://api.telegram.org/file/bot${process.env.BOT_TOKEN}/${sticker.file_path}`;
+    const tgRes = await fetch(telegramUrl);
+    const buffer = await tgRes.arrayBuffer();
+
+    zlib.unzip(Buffer.from(buffer), (err, jsonBuffer) => {
+      if (err) {
+        return res.status(500).json({ error: 'Failed to decode sticker' });
+      }
+
+      try {
+        const json = JSON.parse(jsonBuffer.toString());
+        res.json(json);
+      } catch (e) {
+        res.status(500).json({ error: 'Invalid JSON' });
+      }
+    });
+  } catch (e) {
+    res.status(500).json({ error: 'Internal error' });
+  }
+});
+
+router.get('/sticker/:id/info', async (req, res) => {
+  try {
+    const sticker = await Sticker.findOne(
+      { file_unique_id: req.params.id },
+      { _id: 0, file_unique_id: 1, is_animated: 1, is_video: 1, emoji: 1, set_name: 1 }
+    );
+
+    if (!sticker) {
+      return res.status(404).json({ error: 'Sticker not found' });
+    }
+
+    res.json(sticker);
+  } catch (error) {
+    res.status(500).json({ error: 'Internal server error' });
+  }
+});
+
+router.get('/stickers', async (req, res) => {
+  try {
+    const { set, limit = 50, offset = 0 } = req.query;
+    const query = set ? { set_name: set } : {};
+
+    const stickers = await Sticker.find(query, {
+      file_unique_id: 1,
+      emoji: 1,
+      set_name: 1,
+      is_animated: 1,
+      is_video: 1
+    })
+      .sort({ created_at: -1 })
+      .skip(parseInt(offset))
+      .limit(parseInt(limit));
+
+    res.json(stickers);
+  } catch (error) {
+    res.status(500).json({ error: 'Internal server error' });
+  }
+});
+
+module.exports = router;
